requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
httpx==0.28.1
python-dotenv==1.0.1

# Database
sqlalchemy>=2.0.36
alembic==1.13.1
aiosqlite==0.19.0

# AI
google-generativeai==0.3.2

# Utilities
jinja2==3.1.3
